Move NextUp to the Metadata
